requirements.txt
faiss-cpu
numpy
pdfplumber
sentence-transformers
google-generativeai
python-dotenv